train.py
"""
Train a Random Forest model for formation energy prediction.
Uses JARVIS leaderboard benchmark data with element fraction descriptors.
"""

import os
import numpy as np
import pickle
import pandas as pd
from tqdm import tqdm
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_absolute_error
from jarvis.db.jsonutils import loadjson
from jarvis.ai.descriptors.elemental import get_element_fraction_desc
import zipfile

# --- Setup ---
os.makedirs("Software", exist_ok=True)
os.chdir("Software")
if not os.path.exists("jarvis_leaderboard"):
    os.system("git clone https://github.com/atomgptlab/jarvis_leaderboard.git")
os.system("pip install -q ./jarvis_leaderboard")
os.chdir("..")

# --- Populate data ---
os.system(
    "jarvis_populate_data.py "
    "--benchmark_file AI-SinglePropertyPrediction-formula_energy-ssub-test-mae "
    "--output_path=Out --json_key formula --id_tag id"
)

# --- Load data ---
dataset_info = loadjson("Out/dataset_info.json")
df = pd.read_csv("Out/id_prop.csv", header=None, names=["formula", "form_energy"])
df["id"] = df.index + 1

tqdm.pandas()
df["desc"] = df["formula"].progress_apply(lambda f: get_element_fraction_desc(f))

train_df = df[: dataset_info["n_train"]]
test_df = df[dataset_info["n_train"] :]

# --- Train ---
X_train = np.array(train_df["desc"].tolist())
y_train = train_df["form_energy"].values
X_test = np.array(test_df["desc"].tolist())
y_test = test_df["form_energy"].values

rf = RandomForestRegressor(
    n_estimators=200,
    max_depth=None,
    min_samples_split=2,
    min_samples_leaf=1,
    max_features="sqrt",
    n_jobs=-1,
    random_state=42,
    verbose=1,
)
rf.fit(X_train, y_train)

# --- Evaluate ---
y_pred = rf.predict(X_test)
mae = mean_absolute_error(y_test, y_pred)
print(f"Test MAE: {mae:.6f}")

# --- Save model ---
with open("rf_form_energy_model.pkl", "wb") as f:
    pickle.dump(rf, f)
print("Model saved to rf_form_energy_model.pkl")

# --- Save results ---
results = pd.DataFrame(
    {
        "id": test_df["id"].values,
        "formula": test_df["formula"].values,
        "prediction": y_pred,
        "target": test_df["form_energy"].values,
    }
)
filename = dataset_info["benchmark_file"] + ".csv"
results.to_csv(filename, index=False)

with zipfile.ZipFile(filename + ".zip", "w", zipfile.ZIP_DEFLATED) as zf:
    zf.write(filename, filename)
print(f"Saved {filename} and {filename}.zip")
