requirements.txt
fastapi
uvicorn
scikit-learn
numpy
pandas
jarvis-tools
tqdm
